src/engine/requirements.txt
pandas==2.3.0
azure-functions
azure-storage-blob
sqlalchemy
pyodbc
